[MIG][UPD][ADD] move edi_storage abstract parts into edi_storage_core_oca - Keep Component part into edi_storage_oca; - Move abstract tests into edi_storage_core_oca; - Move data (+ demo) into edi_storage_core_oca; - Move security into edi_storage_core_oca; - Move views into edi_storage_core_oca.
