+++ edi_storage_core_oca/__init__.py
@@ -1,0 +1,1 @@
+from . import models
