+++ edi_storage_core_oca/data/cron.xml
@@ -6,7 +6,7 @@
         <field name="user_id" ref="base.user_root" />
         <field name="interval_number">1</field>
         <field name="interval_type">hours</field>
-        <field name="model_id" ref="edi_oca.model_edi_backend" />
+        <field name="model_id" ref="model_edi_backend" />
         <field name="state">code</field>
         <field
             name="code"
