+++ edi_storage_core_oca/data/job_channel_data.xml
@@ -1,6 +1,6 @@
 <odoo noupdate="1">
     <record id="channel_edi_storage" model="queue.job.channel">
         <field name="name">edi_storage</field>
-        <field name="parent_id" ref="edi_oca.channel_edi_root" />
+        <field name="parent_id" ref="edi_core_oca.channel_edi_root" />
     </record>
 </odoo>
