+++ edi_storage_core_oca/demo/edi_backend_demo.xml
@@ -2,7 +2,7 @@
 <odoo>
     <record id="demo_edi_backend_storage" model="edi.backend">
         <field name="name">Storage Demo EDI backend</field>
-        <field name="backend_type_id" ref="edi_oca.demo_edi_backend_type" />
+        <field name="backend_type_id" ref="edi_core_oca.demo_edi_backend_type" />
         <field name="storage_id" ref="fs_storage.fs_storage_demo" />
         <field name="input_dir_pending">demo_in/pending</field>
         <field name="input_dir_done">demo_in/done</field>
