+++ edi_storage_core_oca/i18n/edi_storage_oca.pot
@@ -1,0 +1,116 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* edi_storage_oca
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 16.0\n"
+"Report-Msgid-Bugs-To: \n"
+"Last-Translator: \n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: edi_storage_oca
+#: model:ir.model,name:edi_storage_oca.model_edi_backend
+msgid "EDI Backend"
+msgstr ""
+
+#. module: edi_storage_oca
+#: model:ir.model,name:edi_storage_oca.model_edi_exchange_type
+msgid "EDI Exchange Type"
+msgstr ""
+
+#. module: edi_storage_oca
+#: model:ir.actions.server,name:edi_storage_oca.cron_check_storage_pending_input_ir_actions_server
+#: model:ir.cron,cron_name:edi_storage_oca.cron_check_storage_pending_input
+msgid "EDI backend storage check pending input"
+msgstr ""
+
+#. module: edi_storage_oca
+#: model:ir.model,name:edi_storage_oca.model_edi_exchange_record
+msgid "EDI exchange Record"
+msgstr ""
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,field_description:edi_storage_oca.field_edi_exchange_type__exchange_filename_pattern
+msgid "Exchange Filename Pattern"
+msgstr ""
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,field_description:edi_storage_oca.field_edi_backend__storage_id
+#: model:ir.model.fields,field_description:edi_storage_oca.field_edi_exchange_record__storage_id
+msgid "FS Storage"
+msgstr ""
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,help:edi_storage_oca.field_edi_exchange_type__exchange_filename_pattern
+msgid ""
+"For output exchange types this should be a formatting string with the following variables available (to be used between brackets, `{}`): `exchange_record`, `record_name`, `type` and `dt`. For instance, a valid string would be {record_name}-{type.code}-{dt}\n"
+"For input exchange types related to storage backends it should be a regex expression to filter the files to be fetched from the pending directory in the related storage. E.g: `.*my-type-[0-9]*.\\.csv`"
+msgstr ""
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,field_description:edi_storage_oca.field_edi_backend__input_dir_done
+msgid "Input done directory"
+msgstr ""
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,field_description:edi_storage_oca.field_edi_backend__input_dir_error
+msgid "Input error directory"
+msgstr ""
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,field_description:edi_storage_oca.field_edi_backend__input_dir_pending
+msgid "Input pending directory"
+msgstr ""
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,field_description:edi_storage_oca.field_edi_backend__output_dir_done
+msgid "Output done directory"
+msgstr ""
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,field_description:edi_storage_oca.field_edi_backend__output_dir_error
+msgid "Output error directory"
+msgstr ""
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,field_description:edi_storage_oca.field_edi_backend__output_dir_pending
+msgid "Output pending directory"
+msgstr ""
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,help:edi_storage_oca.field_edi_backend__input_dir_done
+#: model:ir.model.fields,help:edi_storage_oca.field_edi_backend__output_dir_done
+msgid "Path to folder for doneful operations"
+msgstr ""
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,help:edi_storage_oca.field_edi_backend__input_dir_error
+#: model:ir.model.fields,help:edi_storage_oca.field_edi_backend__output_dir_error
+msgid "Path to folder for error operations"
+msgstr ""
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,help:edi_storage_oca.field_edi_backend__input_dir_pending
+#: model:ir.model.fields,help:edi_storage_oca.field_edi_backend__output_dir_pending
+msgid "Path to folder for pending operations"
+msgstr ""
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,help:edi_storage_oca.field_edi_exchange_record__storage_id
+msgid "Record created from a file found in this FS storage"
+msgstr ""
+
+#. module: edi_storage_oca
+#: model_terms:ir.ui.view,arch_db:edi_storage_oca.edi_backend_view_form
+msgid "Storage"
+msgstr ""
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,help:edi_storage_oca.field_edi_backend__storage_id
+msgid "Storage for in-out files"
+msgstr ""
