+++ edi_storage_core_oca/i18n/it.po
@@ -1,0 +1,126 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* edi_storage_oca
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 16.0\n"
+"Report-Msgid-Bugs-To: \n"
+"PO-Revision-Date: 2024-11-15 17:06+0000\n"
+"Last-Translator: mymage <stefano.consolaro@mymage.it>\n"
+"Language-Team: none\n"
+"Language: it\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: nplurals=2; plural=n != 1;\n"
+"X-Generator: Weblate 5.6.2\n"
+
+#. module: edi_storage_oca
+#: model:ir.model,name:edi_storage_oca.model_edi_backend
+msgid "EDI Backend"
+msgstr "Backend EDI"
+
+#. module: edi_storage_oca
+#: model:ir.model,name:edi_storage_oca.model_edi_exchange_type
+msgid "EDI Exchange Type"
+msgstr "Tipo scambio EDI"
+
+#. module: edi_storage_oca
+#: model:ir.actions.server,name:edi_storage_oca.cron_check_storage_pending_input_ir_actions_server
+#: model:ir.cron,cron_name:edi_storage_oca.cron_check_storage_pending_input
+msgid "EDI backend storage check pending input"
+msgstr "Controllo ingressi in attesa deposito backend EDI"
+
+#. module: edi_storage_oca
+#: model:ir.model,name:edi_storage_oca.model_edi_exchange_record
+msgid "EDI exchange Record"
+msgstr "Record di scambio EDI"
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,field_description:edi_storage_oca.field_edi_exchange_type__exchange_filename_pattern
+msgid "Exchange Filename Pattern"
+msgstr "Schema nome file scambio"
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,field_description:edi_storage_oca.field_edi_backend__storage_id
+#: model:ir.model.fields,field_description:edi_storage_oca.field_edi_exchange_record__storage_id
+msgid "FS Storage"
+msgstr "Deposito FS"
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,help:edi_storage_oca.field_edi_exchange_type__exchange_filename_pattern
+msgid ""
+"For output exchange types this should be a formatting string with the following variables available (to be used between brackets, `{}`): `exchange_record`, `record_name`, `type` and `dt`. For instance, a valid string would be {record_name}-{type.code}-{dt}\n"
+"For input exchange types related to storage backends it should be a regex expression to filter the files to be fetched from the pending directory in the related storage. E.g: `.*my-type-[0-9]*.\\.csv`"
+msgstr ""
+"Per i tipi di scambio in uscita questo dovrebbe essere una stringa di "
+"formattazione con le seguenti variabili disponibili (da utilizzare tra "
+"parentesi, `{}`): `exchange_record`, `record_name`, `type` e `dt`. Per "
+"esempio, una stringa valida sarà {record_name}-{type.code}-{dt}\n"
+"Per tipi scambio relativi ai backend deposito dovrebbe essere una "
+"espressione regolare per filtrare i campi da alimentare dalla cartella in "
+"attesa nel deposito relativo. Es. `.*my-type-[0-9]*.\\.csv`"
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,field_description:edi_storage_oca.field_edi_backend__input_dir_done
+msgid "Input done directory"
+msgstr "Cartella ingressi eseguiti"
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,field_description:edi_storage_oca.field_edi_backend__input_dir_error
+msgid "Input error directory"
+msgstr "Cartella ingressi in errore"
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,field_description:edi_storage_oca.field_edi_backend__input_dir_pending
+msgid "Input pending directory"
+msgstr "Cartella ingressi in attesa"
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,field_description:edi_storage_oca.field_edi_backend__output_dir_done
+msgid "Output done directory"
+msgstr "Cartella uscite eseguite"
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,field_description:edi_storage_oca.field_edi_backend__output_dir_error
+msgid "Output error directory"
+msgstr "Cartella uscite in errore"
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,field_description:edi_storage_oca.field_edi_backend__output_dir_pending
+msgid "Output pending directory"
+msgstr "Cartella uscite in attesa"
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,help:edi_storage_oca.field_edi_backend__input_dir_done
+#: model:ir.model.fields,help:edi_storage_oca.field_edi_backend__output_dir_done
+msgid "Path to folder for doneful operations"
+msgstr "Percorso alla cartella per le operazioni completamente eseguite"
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,help:edi_storage_oca.field_edi_backend__input_dir_error
+#: model:ir.model.fields,help:edi_storage_oca.field_edi_backend__output_dir_error
+msgid "Path to folder for error operations"
+msgstr "Percorso alla cartella per le operazioni in errore"
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,help:edi_storage_oca.field_edi_backend__input_dir_pending
+#: model:ir.model.fields,help:edi_storage_oca.field_edi_backend__output_dir_pending
+msgid "Path to folder for pending operations"
+msgstr "Percorso alla cartella per le operazioni in attesa"
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,help:edi_storage_oca.field_edi_exchange_record__storage_id
+msgid "Record created from a file found in this FS storage"
+msgstr "Record creato da un file trovato nel deposito FS"
+
+#. module: edi_storage_oca
+#: model_terms:ir.ui.view,arch_db:edi_storage_oca.edi_backend_view_form
+msgid "Storage"
+msgstr "Deposito"
+
+#. module: edi_storage_oca
+#: model:ir.model.fields,help:edi_storage_oca.field_edi_backend__storage_id
+msgid "Storage for in-out files"
+msgstr "Deposito per i file ingresso-uscita"
