+++ edi_storage_core_oca/models/__init__.py
@@ -1,0 +1,3 @@
+from . import edi_backend
+from . import edi_exchange_type
+from . import edi_exchange_record
