+++ edi_storage_core_oca/models/edi_exchange_type.py
