+++ edi_storage_core_oca/readme/CONTRIBUTORS.md
@@ -1,0 +1,4 @@
+- Simone Orsi \<simahawk@gmail.com\>
+- Foram Shah \<foram.shah@initos.com\>
+- Lois Rilo \<lois.rilo@forgeflow.com\>
+- Duong (Tran Quoc) \<duongtq@trobz.com\>
