+++ edi_storage_core_oca/readme/CREDITS.md
@@ -1,0 +1,2 @@
+The migration of this module from 15.0 to 16.0 was financially supported
+by Camptocamp.
