+++ edi_storage_core_oca/readme/ROADMAP.md
@@ -1,0 +1,1 @@
+- clean deprecated methods in the storage
