+++ edi_storage_core_oca/readme/USAGE.md
@@ -1,0 +1,2 @@
+Go to "EDI -\> EDI backend" then configure your backend to use a storage
+backend.
