+++ edi_storage_core_oca/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_exchange_type
