+++ edi_storage_core_oca/tests/test_exchange_type.py
@@ -1,7 +1,7 @@
 # Copyright 2022 Camptocamp SA (https://www.camptocamp.com).
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 
-from odoo.addons.edi_oca.tests.common import EDIBackendCommonTestCase
+from odoo.addons.edi_core_oca.tests.common import EDIBackendCommonTestCase
 
 
 class EDIExchangeTypeTestCase(EDIBackendCommonTestCase):
