+++ edi_storage_core_oca/views/edi_backend_views.xml
@@ -2,7 +2,7 @@
 <odoo>
     <record id="edi_backend_view_form" model="ir.ui.view">
         <field name="model">edi.backend</field>
-        <field name="inherit_id" ref="edi_oca.edi_backend_view_form" />
+        <field name="inherit_id" ref="edi_core_oca.edi_backend_view_form" />
         <field name="arch" type="xml">
             <notebook position="inside">
                 <page name="storage_config" string="Storage">
