+++ edi_storage_oca/components/send.py
@@ -4,8 +4,7 @@
 
 
 from odoo.addons.component.core import Component
-
-from .. import utils
+from odoo.addons.edi_storage_core_oca import utils
 
 
 class EDIStorageSendComponent(Component):
