+++ edi_storage_oca/models/__init__.py
@@ -1,3 +1,1 @@
 from . import edi_backend
-from . import edi_exchange_type
-from . import edi_exchange_record
