+++ edi_storage_oca/tests/__init__.py
@@ -2,4 +2,3 @@ from . import test_edi_backend_storage
 from . import test_components_base
 from . import test_component_match
 from . import test_edi_storage_listener
-from . import test_exchange_type
