+++ edi_storage_oca/tests/test_component_match.py
@@ -14,7 +14,7 @@ class EDIBackendTestCase(EDIBackendCommonComponentRegistryTestCase):
 
     @classmethod
     def _get_backend(cls):
-        return cls.env.ref("edi_storage_oca.demo_edi_backend_storage")
+        return cls.env.ref("edi_storage_core_oca.demo_edi_backend_storage")
 
     def test_component_match(self):
         """Lookup with special match method."""
